@@ -2,9 +2,8 @@
  * Copyright 2018 Ambrosus Inc.
  * Email: tech@ambrosus.com
  */
-import React, { Component } from 'react';
+import React, { useState, useRef } from 'react';
 import { formatDate } from '../../utils';
-import ReactDOM from 'react-dom';
 import Calendar from 'react-calendar';
 
 import './CalendarInput.scss';
@@ -12,74 +11,65 @@ import { ICalendarInput } from '../../../interfaces';
 
 import { Input } from '../';
 
-export default class CalendarInput extends Component<ICalendarInput, any> {
-    public state: any = {
-        otherProps: {},
-        error: '',
-        calendar: false,
-        date: '',
-    };
-    private calendarRef: any;
+function CalendarInput(props: ICalendarInput) {
+    const [calendar, setCalendar] = useState(false);
+    const [date, setDate] = useState('');
+    const calendarRef: any = useRef(null);
 
-    constructor(props: ICalendarInput) {
-        super(props);
-    }
+    const { daterange, value, onChange, onFocus, onBlur, ...otherProps } = props;
 
-    private closeOnScroll = (event: Event) => {
-        this.closeCalendar();
-    }
-
-    private onChoose = (date: Date | Date[]) => {
-        if (!!this.props.daterange) {
-            this.setState({
-                date: `${formatDate(new Date(date[0].toString()))} - ${formatDate(new Date(date[1].toString()))}`,
-            });
+    const onChoose = (_date: Date | Date[]) => {
+        if (!!daterange) {
+            setDate(`${formatDate(new Date(_date[0].toString()))} - ${formatDate(new Date(_date[1].toString()))}`);
         } else {
-            this.setState({ date: formatDate(new Date(date.toString())) });
+            setDate(formatDate(new Date(_date.toString())));
         }
-        this.closeCalendar();
-    }
+        closeCalendar();
+    };
 
-    private onChange = (event: any) => {
-        this.setState({ date: event.target.value });
-        if (this.props.changed) {
-            this.props.changed(event);
+    const _onChange = (event: any) => {
+        setDate(event.target.value);
+        if (onChange) {
+            onChange(event);
         }
-    }
+    };
 
-    private handleClickOutsideCalendar = (event: any) => {
-        if (this.calendarRef && !this.calendarRef.contains(event.target) && !event.target.classList.contains('react-calendar__tile')) {
-            this.closeCalendar();
+    const handleClickOutsideCalendar = (event: any) => {
+        if (
+            calendarRef &&
+            calendarRef.current &&
+            !calendarRef.current.contains(event.target) &&
+            !event.target.classList.contains('react-calendar__tile')
+        ) {
+            closeCalendar();
         }
-    }
+    };
 
-    private openCalendar = () => {
-        this.setState({ calendar: true });
-        document.addEventListener('click', this.handleClickOutsideCalendar);
-    }
+    const openCalendar = () => {
+        setCalendar(true);
+        document.addEventListener('click', handleClickOutsideCalendar);
+    };
 
-    private closeCalendar = () => {
-        this.setState({ calendar: false });
-        document.removeEventListener('click', this.handleClickOutsideCalendar);
-    }
+    const closeCalendar = () => {
+        setCalendar(false);
+        document.removeEventListener('click', handleClickOutsideCalendar);
+    };
 
-    public render() {
-        const { value, changed, ...otherProps } = this.props;
+    const _placeholder = formatDate(new Date());
 
-        const _placeholder = formatDate(new Date());
+    const _value = date;
 
-        const _value = this.state.date;
-
-        return (
-            <div className='AMB-CalendarInput'>
-                <Input {...otherProps} value={_value} placeholder={_placeholder} changed={this.onChange}
-                    onClick={() => !otherProps.disabled && this.openCalendar()} />
-                {this.state.calendar && (
-                    <div className='calendar' ref={r => this.calendarRef = r}>
-                        <Calendar onChange={this.onChoose} selectRange={!!this.props.daterange} />
-                    </div>
-                )}
-            </div>
-        );
-    }
+    return (
+        <div className='AMB-CalendarInput'>
+            <Input {...otherProps} value={_value} placeholder={_placeholder} onChange={_onChange}
+                onClick={() => !otherProps.disabled && openCalendar()} />
+            {calendar && (
+                <div className='calendar' ref={calendarRef}>
+                    <Calendar onChange={onChoose} selectRange={!!daterange} />
+                </div>
+            )}
+        </div>
+    );
 }
+
+export default CalendarInput;
