@@ -2,7 +2,7 @@
  * Copyright 2018 Ambrosus Inc.
  * Email: tech@ambrosus.com
  */
-import React from 'react';
+import React, { useState, useCallback, useRef } from 'react';
 import { SVG } from '../../utils';
 
 import './Textarea.scss';
@@ -11,19 +11,23 @@ import { ITextarea } from '../../../interfaces';
 import iconSuccess from '../../../assets/svg/success.svg';
 import iconInfo from '../../../assets/svg/info.svg';
 
-const Textarea = (props: ITextarea) => {
-    const { label, className, value, changed, invalid, shouldValidate, placeholder, children,
-        touched, errors, disabled, light, info, ...otherProps } = props;
+function Textarea(props: ITextarea) {
+    const [touched, setTouched] = useState(false);
+    const inputRef: any = useRef(null);
+
+    const onFocus = useCallback(() => setTouched(true), [touched]);
+
+    const { label, className, value, onChange, children, error, disabled, light, info, ...otherProps } = props;
 
     const classes: any = [
         'AMB-Textarea',
         `${light && 'light' || ''}`,
         `${disabled && 'disabled' || ''}`,
-        `${touched && !invalid && 'valid' || ''}`,
+        `${touched && !error && 'valid' || ''}`,
         `${className || ''}`.trim(),
     ].filter(Boolean);
 
-    if (invalid && shouldValidate && touched) {
+    if (error && touched) {
         classes.push('error');
     }
 
@@ -34,7 +38,9 @@ const Textarea = (props: ITextarea) => {
             <div className='title'>
                 <span>{label}</span>
                 <div className='meta'>
-                    {touched && !invalid && <SVG className='SVG' src={iconSuccess} />}
+                    {
+                        touched && !error && !!String(inputRef && inputRef.current && inputRef.current.value).trim() && <SVG className='SVG' src={iconSuccess} />
+                    }
                     {info && (
                         <div className='info'>
                             <SVG className='SVG' src={iconInfo} />
@@ -45,17 +51,18 @@ const Textarea = (props: ITextarea) => {
             </div>
             <div className='textarea'>
                 <textarea
+                    ref={inputRef}
                     value={_value}
-                    onChange={changed}
+                    onChange={onChange}
+                    onFocus={onFocus}
                     spellCheck={false}
-                    placeholder={placeholder}
                     disabled={disabled}
                 ></textarea>
                 <div className='border'></div>
             </div>
-            {invalid && shouldValidate && touched && <p className='message'>{errors}</p>}
+            {error && touched && <p className='message'>{error}</p>}
         </label>
     );
-};
+}
 
 export default Textarea;
